Abstract similar `SetDiff`-based nested blocks into `BlockSetAttributeDefinition` interface
I've moved some of the `ServiceAttributeDefinition` code into `fastly/block.go`.
This interface represents any nested block, regardless of how it's implemented, and it is this interface that talks to the base service code.

Underneath this, most (but not all) blocks that are of `TypeSet` use a very similar structure based on the `SetDiff` "library" function.
To simplify this, and reduce duplicated code, I have added a new `BlockSetAttributeDefinition` interface for these common set-based blocks.
This interface contains CRUD functions and a `GetSchema` function so that these files can look similar to full on resource files.
There is then a common implementation of the `ServiceAttributeDefinition` interface for the `BlockSetAttributeDefinition` implementations which handles adapting the interface and calling `SetDiff`.
This code is in `fastly/block_set.go`.

To illustrate how these interfaces will interact, I have manually updated about three of the blocks (first few alphabetically) into the new format.